threadpool/src/job_orchestrator/handlers/job1_handler.py
from job_orchestrator.handlers.generic_job_handler import GenericJobHandler

"""
This module defines the Job1Handler class, a specialized extension of the GenericJobHandler designed to manage
the execution lifecycle of 'Job 1'. The Job1Handler class customizes the preparation, execution, and finalization
phases of the job to cater to specific requirements and configurations of 'Job 1'.

The Job1Handler is particularly useful in scenarios where 'Job 1' requires distinct setup, execution logic, or
cleanup processes that differ from the standard job handling provided by GenericJobHandler. This class ensures
that these specific needs are addressed effectively while maintaining the overall structure and workflow defined
by the GenericJobHandler.

Classes:
    Job1Handler: Manages the execution lifecycle of 'Job 1', customizing the preparation, execution, and cleanup phases.

Example usage:
    # Assuming the necessary task configurations for 'Job 1' are defined.
    job_handler = Job1Handler()
    tasks = [
        {"name": "task1", "details": "Specific details for Job 1"},
        {"name": "task2", "details": "More details for Job 1"}
    ]
    job_handler.before_job()  # Prepare the environment for 'Job 1'
    job_handler.execute_tasks(tasks)  # Execute the tasks for 'Job 1'
    job_handler.after_job()  # Cleanup and finalize after 'Job 1'
"""

class Job1Handler(GenericJobHandler):
    """
    A specialized job handler for 'Job 1' that extends the GenericJobHandler.
    """
    def before_job(self):
        """
        Prepare the environment or settings specific to 'Job 1'.
        """
        print("Preparing Job 1...")

    def execute_tasks(self, tasks):
        """
        Execute the tasks specifically tailored for 'Job 1'.
        """
        # Custom task execution logic for Job 1 can be placed here.
        super().execute_tasks(tasks)  

    def after_job(self):
        """
        Handle cleanup and finalization after 'Job 1' has been executed.
        """
        print("Job 1 completed successfully.")
